requirements.txt
accelerate==1.6.0
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
backoff==2.2.1
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
defusedxml==0.7.1
distro==1.9.0
filelock==3.18.0
fsspec==2025.3.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.30.1
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core>=0.3.58,<0.4.0
langchain-anthropic==0.3.3
langchain-ollama==0.2.2
langchain-openai==0.3.1
langsmith==0.3.24
markdownify==0.14.1
MarkupSafe==3.0.2
monotonic==1.6
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
ollama==0.4.7
openai==1.70.0
orjson==3.10.16
packaging==24.2
pillow==11.1.0
playwright==1.51.0
posthog==3.23.0
psutil==7.0.0
pydantic==2.11.2
pydantic_core==2.33.1
pyee==12.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.0.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sympy==1.13.1
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.51.0
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
zstandard==0.23.0
flask==3.1.0
flask_cors==5.0.1
mem0ai==0.1.96
faiss-cpu==1.11.0
screeninfo==0.8.1
mcp==1.7.1
flask[async]
langchain==0.3.25
html_sanitizer==2.5.0
